Move help message generator to Launcher
